attempting to fix windows build memory error

+++ csrc/cpu/radius_cpu.h
@@ -7,10 +7,10 @@
 #include "compat.h"
 
 torch::Tensor radius_cpu(torch::Tensor query, torch::Tensor support,
-			 			 float radius, int max_num);
+			 			 double radius, int64_t max_num);
 
 torch::Tensor batch_radius_cpu(torch::Tensor query,
 			       torch::Tensor support,
 			       torch::Tensor query_batch,
 			       torch::Tensor support_batch,
-			       float radius, int max_num);
+			       double radius, int64_t max_num);
