+++ csrc/cpu/utils/neighbors.h
@@ -11,7 +11,7 @@ using namespace std;
 
 template<typename scalar_t>
 int nanoflann_neighbors(vector<scalar_t>& queries, vector<scalar_t>& supports,
-			vector<long>& neighbors_indices, float radius, int dim, int max_num);
+			vector<long>& neighbors_indices, double radius, int dim, int64_t max_num);
 
 template<typename scalar_t>
 int batch_nanoflann_neighbors (vector<scalar_t>& queries,
@@ -19,4 +19,4 @@ int batch_nanoflann_neighbors (vector<scalar_t>& queries,
                                vector<long>& q_batches,
                                vector<long>& s_batches,
                                vector<long>& neighbors_indices,
-                               float radius, int dim, int max_num);
+                               double radius, int dim, int64_t max_num);
